pyproject.toml
[tool.poetry]
name = "tabtab"
version = "0.1.0"
description = ""
authors = ["Misha Behersky <bmwant@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-telegram-bot = ">=12"
attrs = "17.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.3.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
